break down interface_parser into a submodule of smaller parts
